backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose[cryptography]
passlib[argon2]
python-multipart
python-dotenv
alembic
pymysql
slowapi
